requirements.txt
numpy
trimesh
pyrender
scipy
transforms3d
open3d
rerun-sdk